+++ changelog.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2.1.15
+
+- Add Walls Reach
+- Add more API
+- Add bug fix for [Not working well with Doors on Diagonal walls](https://github.com/p4535992/foundryvtt-arms-reach/issues/40)
+
 ## 2.1.14
 
 - little update for the reset fog and door feature